# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Full API DTCL",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/blog/status/{id}": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "status Blog",
                "description": "status Blog",
                "operationId": "statusBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/user": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get all user",
                "operationId": "someRoute",
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/user/{id}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get user",
                "operationId": "isUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "get user"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Delete user",
                "description": "Delete user",
                "operationId": "DeleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/user/band/{id}": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Band user",
                "description": "Band user",
                "operationId": "BandUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/user/active/{id}": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Active user",
                "description": "Active user",
                "operationId": "ActiveUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "get all Blog",
                "operationId": "Blog",
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "add Blog",
                "operationId": "addBlog",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "text"
                                    },
                                    "image[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blog/{id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "detail Blog",
                "operationId": "detailBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "blog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                }
            },
            "put": {
                "tags": [
                    "Blog"
                ],
                "summary": "add Blog",
                "operationId": "updateBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "blog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "content",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blog"
                ],
                "summary": "delete Blog",
                "operationId": "deleteBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "blog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blog/cmt/{id}": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "add Blog",
                "operationId": "cmtBlog",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "blog_id": {
                                        "type": "integer"
                                    },
                                    "content": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}